Refactoring: move and split DotcomponentsDataModel.scala
