+++ article/app/renderers/RemoteRenderer.scala
@@ -7,7 +7,7 @@ import concurrent.CircuitBreakerRegistry
 import conf.Configuration
 import conf.switches.Switches.CircuitBreakerSwitch
 import model.Cached.RevalidatableResult
-import model.dotcomrendering.{DCRDataModel, DotcomponentsDataModel}
+import model.dotcomrendering.{DCRDataModel, DotcomRenderingTransforms}
 import model.{Cached, NoCache, PageWithStoryPackage}
 import play.api.libs.ws.{WSClient, WSResponse}
 import play.api.mvc.{RequestHeader, Result}
@@ -56,7 +56,7 @@ class RemoteRenderer extends Logging {
       blocks: Blocks,
       pageType: PageType,
   )(implicit request: RequestHeader): Future[Result] = {
-    val dataModel = DotcomponentsDataModel.fromArticle(page, request, blocks, pageType)
+    val dataModel = DotcomRenderingTransforms.fromArticle(page, request, blocks, pageType)
     val json = DCRDataModel.toJson(dataModel)
 
     def handler(response: WSResponse): Result = {
@@ -88,7 +88,7 @@ class RemoteRenderer extends Logging {
       blocks: Blocks,
       pageType: PageType,
   )(implicit request: RequestHeader): Future[Result] = {
-    val dataModel = DotcomponentsDataModel.fromArticle(page, request, blocks, pageType)
+    val dataModel = DotcomRenderingTransforms.fromArticle(page, request, blocks, pageType)
     val json = DCRDataModel.toJson(dataModel)
 
     def handler(response: WSResponse): Result = {
