+++ common/app/model/dotcomrendering/DotcomRenderingTransforms.scala
@@ -39,303 +39,7 @@ import org.joda.time.DateTime
 import org.joda.time.format.DateTimeFormat
 import views.support.JavaScriptPage
 
-// We have introduced our own set of objects for serializing data to the DotComponents API,
-// because we don't want people changing the core frontend models and as a side effect,
-// making them incompatible with Dotcomponents. By having our own set of models, there's
-// only one reason for change.
-// exceptions: we do resuse the existing Nav & BlockElement classes right now
-
-// -----------------------------------------------------------------
-// Supporting Types
-// -----------------------------------------------------------------
-
-case class ArticlePage(article: Article, related: RelatedContent) extends PageWithStoryPackage
-
-case class Tag(
-    id: String,
-    `type`: String,
-    title: String,
-    twitterHandle: Option[String],
-    bylineImageUrl: Option[String],
-)
-
-object Tag {
-  implicit val writes = Json.writes[Tag]
-}
-
-case class Block(
-    id: String,
-    elements: List[PageElement],
-    createdOn: Option[Long],
-    createdOnDisplay: Option[String],
-    lastUpdated: Option[Long],
-    lastUpdatedDisplay: Option[String],
-    firstPublished: Option[Long],
-    firstPublishedDisplay: Option[String],
-    title: Option[String],
-    primaryDateLine: String,
-    secondaryDateLine: String,
-)
-
-object Block {
-  implicit val pageElementWrites = PageElement.pageElementWrites
-  implicit val writes = Json.writes[Block]
-}
-
-case class Pagination(
-    currentPage: Int,
-    totalPages: Int,
-    newest: Option[String],
-    newer: Option[String],
-    oldest: Option[String],
-    older: Option[String],
-)
-
-object Pagination {
-  implicit val writes = Json.writes[Pagination]
-}
-
-case class ReaderRevenueLink(
-    contribute: String,
-    subscribe: String,
-    support: String,
-)
-
-object ReaderRevenueLink {
-  implicit val writes = Json.writes[ReaderRevenueLink]
-}
-
-case class ReaderRevenueLinks(
-    header: ReaderRevenueLink,
-    footer: ReaderRevenueLink,
-    sideMenu: ReaderRevenueLink,
-    ampHeader: ReaderRevenueLink,
-    ampFooter: ReaderRevenueLink,
-)
-
-object ReaderRevenueLinks {
-  implicit val writes = Json.writes[ReaderRevenueLinks]
-}
-
-case class Commercial(
-    editionCommercialProperties: Map[String, EditionCommercialProperties],
-    prebidIndexSites: List[PrebidIndexSite],
-    commercialProperties: Option[CommercialProperties],
-    pageType: PageType,
-)
-
-object Commercial {
-  implicit val writes = Json.writes[Commercial]
-}
-
-case class Config(
-    switches: Map[String, Boolean],
-    abTests: Map[String, String],
-    commercialBundleUrl: String,
-    googletagUrl: String,
-    stage: String,
-    frontendAssetsFullURL: String,
-    ampIframeUrl: String,
-)
-
-object Config {
-  implicit val writes = Json.writes[Config]
-}
-
-case class SubMetaLink(
-    url: String,
-    title: String,
-)
-
-object SubMetaLink {
-  implicit val format = Json.format[SubMetaLink]
-
-  def apply(sml: model.SubMetaLink): SubMetaLink = {
-    SubMetaLink(
-      url = sml.link,
-      title = sml.text,
-    )
-  }
-}
-
-case class Author(
-    byline: Option[String],
-    twitterHandle: Option[String],
-)
-
-object Author {
-  implicit val writes = Json.writes[Author]
-}
-
-case class DCRBadge(seriesTag: String, imageUrl: String)
-
-object DCRBadge {
-  implicit val writes = Json.writes[DCRBadge]
-}
-
-case class Nav(
-    currentUrl: String,
-    pillars: Seq[NavLink],
-    otherLinks: Seq[NavLink],
-    brandExtensions: Seq[NavLink],
-    currentNavLink: Option[NavLink],
-    currentParent: Option[NavLink],
-    currentPillar: Option[NavLink],
-    subNavSections: Option[Subnav],
-    readerRevenueLinks: ReaderRevenueLinks,
-)
-
-object Nav {
-  implicit val navlinkWrites = Json.writes[NavLink]
-  implicit val flatSubnavWrites = Json.writes[FlatSubnav]
-  implicit val parentSubnavWrites = Json.writes[ParentSubnav]
-  implicit val subnavWrites = Writes[Subnav] {
-    case nav: FlatSubnav   => flatSubnavWrites.writes(nav)
-    case nav: ParentSubnav => parentSubnavWrites.writes(nav)
-  }
-  implicit val writes = Json.writes[Nav]
-}
-
-case class PageFooter(
-    footerLinks: Seq[Seq[FooterLink]],
-)
-
-object PageFooter {
-  implicit val footerLinkWrites: Writes[FooterLink] = Json.writes[FooterLink]
-  implicit val writes = Json.writes[PageFooter]
-}
-
-// -----------------------------------------------------------------
-// DCR DataModel
-// -----------------------------------------------------------------
-
-case class DCRDataModel(
-    version: Int,
-    headline: String,
-    standfirst: String,
-    webTitle: String,
-    mainMediaElements: List[PageElement],
-    main: String,
-    keyEvents: List[Block],
-    blocks: List[Block],
-    pagination: Option[Pagination],
-    author: Author,
-    webPublicationDate: String,
-    webPublicationDateDisplay: String, // TODO remove
-    editionLongForm: String,
-    editionId: String,
-    pageId: String,
-    tags: List[Tag],
-    pillar: String,
-    isImmersive: Boolean,
-    sectionLabel: String,
-    sectionUrl: String,
-    sectionName: Option[String],
-    subMetaSectionLinks: List[SubMetaLink],
-    subMetaKeywordLinks: List[SubMetaLink],
-    shouldHideAds: Boolean,
-    isAdFreeUser: Boolean,
-    webURL: String,
-    linkedData: List[LinkedData],
-    openGraphData: Map[String, String],
-    twitterData: Map[String, String],
-    config: JsObject,
-    guardianBaseURL: String,
-    contentType: String,
-    hasRelated: Boolean,
-    hasStoryPackage: Boolean,
-    beaconURL: String,
-    isCommentable: Boolean,
-    commercialProperties: Map[String, EditionCommercialProperties],
-    pageType: PageType,
-    starRating: Option[Int],
-    trailText: String,
-    nav: Nav,
-    designType: String,
-    showBottomSocialButtons: Boolean,
-    pageFooter: PageFooter,
-    publication: String,
-    shouldHideReaderRevenue: Boolean,
-    // slot machine (temporary for contributions development)
-    slotMachineFlags: String,
-    contributionsServiceUrl: String,
-    badge: Option[DCRBadge],
-    // Match Data
-    matchUrl: Option[String], // Optional url used for match data
-)
-
-object DCRDataModel {
-
-  implicit val pageElementWrites = PageElement.pageElementWrites
-
-  implicit val writes = new Writes[DCRDataModel] {
-    def writes(model: DCRDataModel) =
-      Json.obj(
-        "version" -> model.version,
-        "headline" -> model.headline,
-        "standfirst" -> model.standfirst,
-        "webTitle" -> model.webTitle,
-        "mainMediaElements" -> Json.toJson(model.mainMediaElements),
-        "main" -> model.main,
-        "keyEvents" -> model.keyEvents,
-        "blocks" -> model.blocks,
-        "pagination" -> model.pagination,
-        "author" -> model.author,
-        "webPublicationDate" -> model.webPublicationDate,
-        "webPublicationDateDisplay" -> model.webPublicationDateDisplay,
-        "editionLongForm" -> model.editionLongForm,
-        "editionId" -> model.editionId,
-        "pageId" -> model.pageId,
-        "tags" -> model.tags,
-        "pillar" -> model.pillar,
-        "isImmersive" -> model.isImmersive,
-        "sectionLabel" -> model.sectionLabel,
-        "sectionUrl" -> model.sectionUrl,
-        "sectionName" -> model.sectionName,
-        "subMetaSectionLinks" -> model.subMetaSectionLinks,
-        "subMetaKeywordLinks" -> model.subMetaKeywordLinks,
-        "shouldHideAds" -> model.shouldHideAds,
-        "isAdFreeUser" -> model.isAdFreeUser,
-        "webURL" -> model.webURL,
-        "linkedData" -> model.linkedData,
-        "openGraphData" -> model.openGraphData,
-        "twitterData" -> model.twitterData,
-        "config" -> model.config,
-        "guardianBaseURL" -> model.guardianBaseURL,
-        "contentType" -> model.contentType,
-        "hasRelated" -> model.hasRelated,
-        "hasStoryPackage" -> model.hasStoryPackage,
-        "beaconURL" -> model.beaconURL,
-        "isCommentable" -> model.isCommentable,
-        "commercialProperties" -> model.commercialProperties,
-        "pageType" -> model.pageType,
-        "starRating" -> model.starRating,
-        "trailText" -> model.trailText,
-        "nav" -> model.nav,
-        "designType" -> model.designType,
-        "showBottomSocialButtons" -> model.showBottomSocialButtons,
-        "pageFooter" -> model.pageFooter,
-        "publication" -> model.publication,
-        "shouldHideReaderRevenue" -> model.shouldHideReaderRevenue,
-        "slotMachineFlags" -> model.slotMachineFlags,
-        "contributionsServiceUrl" -> model.contributionsServiceUrl,
-        "badge" -> model.badge,
-        "matchUrl" -> model.matchUrl,
-      )
-  }
-
-  def toJson(model: DCRDataModel): String = {
-    def withoutNull(json: JsValue): JsValue =
-      json match {
-        case JsObject(fields) => JsObject(fields.filterNot { case (_, value) => value == JsNull })
-        case other            => other
-      }
-    val jsValue = Json.toJson(model)
-    Json.stringify(withoutNull(jsValue))
-  }
-}
-
-object DotcomponentsDataModel {
+object DotcomRenderingTransforms {
 
   private def makeMatchUrl(articlePage: PageWithStoryPackage): Option[String] = {
